Update README.md
Took some examples from the cart package so members can stay here and do not have to go to the cart package itself. It is your own text from the cart package by the way. When i have something to add in the track i am looking at the inner workings of the package i will commit my information on that document ofcourse.